Update folder structure, change name of the folders, break main.dart and create app.dart  and update some code.
