+++ pubspec.yaml
@@ -1,5 +1,5 @@
 name: ribbitapp
-description: 'A new Flutter project.'
+description: 'A flutter project to create a crypto wallet.'
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
